+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+- Fix cannot find the type for custom attribute error. [#21](https://github.com/BernieWhite/PSRule.Rules.Azure/issues/21)
+
 ## v0.1.0-B190536
 
 - Initial pre-release